pyproject.toml
[tool.poetry]
name = "dbt-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
duckdb = "^1.0.0"
pandas = "^2.2.2"
pyarrow = "^16.1.0"
fire = "^0.6.0"
loguru = "^0.7.2"
pydantic = "^2.7.4"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"
dbt-duckdb = "^1.8.1"
snowflake-snowpark-python = "^1.19.0"
dbt-snowflake = "^1.8.3"
seaborn = "^0.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
ruff = "^0.4.10"
shandy-sqlfmt = "^0.21.3"

[tool.sqlfmt]
exclude=["target/**/*", "dbt_packages/**/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

package-mode = false